feat(hyprland): add pyprland to reload script

+++ hypr/.config/hypr/scripts/reload_config.sh
@@ -2,7 +2,17 @@
 
 waypaper --restore
 
-pkill waybar
+# Restart waybar if running
+if pgrep -x waybar >/dev/null; then
+  pkill waybar
+fi
 waybar &
 
+# Restart pyprland if running
+if pgrep -x pypr >/dev/null; then
+  pkill pypr
+fi
+pypr &
+
+# Apply spicetify
 spicetify apply
